+++ private/buf/cmd/buf/command/registry/module/moduleinfo/moduleinfo.go
@@ -12,34 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package repositorydeprecate
+package moduleinfo
 
 import (
 	"context"
 	"fmt"
 
+	modulev1 "buf.build/gen/go/bufbuild/registry/protocolbuffers/go/buf/registry/module/v1"
 	"connectrpc.com/connect"
 	"github.com/bufbuild/buf/private/buf/bufcli"
+	"github.com/bufbuild/buf/private/buf/bufprint"
+	"github.com/bufbuild/buf/private/bufpkg/bufapi"
 	"github.com/bufbuild/buf/private/bufpkg/bufmodule"
-	"github.com/bufbuild/buf/private/gen/proto/connect/buf/alpha/registry/v1alpha1/registryv1alpha1connect"
-	registryv1alpha1 "github.com/bufbuild/buf/private/gen/proto/go/buf/alpha/registry/v1alpha1"
 	"github.com/bufbuild/buf/private/pkg/app/appcmd"
 	"github.com/bufbuild/buf/private/pkg/app/appext"
-	"github.com/bufbuild/buf/private/pkg/connectclient"
 	"github.com/bufbuild/buf/private/pkg/syserror"
 	"github.com/spf13/pflag"
 )
 
-const (
-	messageFlagName = "message"
-)
+const formatFlagName = "format"
 
 // NewCommand returns a new Command
-func NewCommand(name string, builder appext.SubCommandBuilder) *appcmd.Command {
+func NewCommand(
+	name string,
+	builder appext.SubCommandBuilder,
+) *appcmd.Command {
 	flags := newFlags()
 	return &appcmd.Command{
-		Use:   name + " <buf.build/owner/repository>",
-		Short: "Deprecate a BSR repository",
+		Use:   name + " <remote/owner/module>",
+		Short: "Get a BSR module",
 		Args:  appcmd.ExactArgs(1),
 		Run: builder.NewRunFunc(
 			func(ctx context.Context, container appext.Container) error {
@@ -51,7 +52,7 @@ func NewCommand(name string, builder appext.SubCommandBuilder) *appcmd.Command {
 }
 
 type flags struct {
-	Message string
+	Format string
 }
 
 func newFlags() *flags {
@@ -60,45 +61,62 @@ func newFlags() *flags {
 
 func (f *flags) Bind(flagSet *pflag.FlagSet) {
 	flagSet.StringVar(
-		&f.Message,
-		messageFlagName,
-		"",
-		`The message to display with deprecation warnings`,
+		&f.Format,
+		formatFlagName,
+		bufprint.FormatText.String(),
+		fmt.Sprintf(`The output format to use. Must be one of %s`, bufprint.AllFormatsString),
 	)
 }
 
-func run(ctx context.Context, container appext.Container, flags *flags) error {
-	bufcli.WarnBetaCommand(ctx, container)
+func run(
+	ctx context.Context,
+	container appext.Container,
+	flags *flags,
+) error {
 	moduleFullName, err := bufmodule.ParseModuleFullName(container.Arg(0))
 	if err != nil {
 		return appcmd.NewInvalidArgumentError(err.Error())
 	}
+	format, err := bufprint.ParseFormat(flags.Format)
+	if err != nil {
+		return appcmd.NewInvalidArgumentError(err.Error())
+	}
+
 	clientConfig, err := bufcli.NewConnectClientConfig(container)
 	if err != nil {
 		return err
 	}
-	service := connectclient.Make(
-		clientConfig,
-		moduleFullName.Registry(),
-		registryv1alpha1connect.NewRepositoryServiceClient,
-	)
-	if _, err := service.DeprecateRepositoryByName(
+	moduleServiceClient := bufapi.NewClientProvider(clientConfig).V1ModuleServiceClient(moduleFullName.Registry())
+	resp, err := moduleServiceClient.GetModules(
 		ctx,
 		connect.NewRequest(
-			&registryv1alpha1.DeprecateRepositoryByNameRequest{
-				OwnerName:          moduleFullName.Owner(),
-				RepositoryName:     moduleFullName.Name(),
-				DeprecationMessage: flags.Message,
+			&modulev1.GetModulesRequest{
+				ModuleRefs: []*modulev1.ModuleRef{
+					{
+						Value: &modulev1.ModuleRef_Name_{
+							Name: &modulev1.ModuleRef_Name{
+								Owner:  moduleFullName.Owner(),
+								Module: moduleFullName.Name(),
+							},
+						},
+					},
+				},
 			},
 		),
-	); err != nil {
+	)
+	if err != nil {
 		if connect.CodeOf(err) == connect.CodeNotFound {
-			return bufcli.NewRepositoryNotFoundError(container.Arg(0))
+			return bufcli.NewModuleNotFoundError(container.Arg(0))
 		}
 		return err
 	}
-	if _, err := fmt.Fprintln(container.Stdout(), "Repository deprecated."); err != nil {
-		return syserror.Wrap(err)
+	modules := resp.Msg.Modules
+	if len(modules) != 1 {
+		return syserror.Newf("unexpected nubmer of modules returned from server: %d", len(modules))
 	}
-	return nil
+	return bufprint.NewModulePrinter(
+		clientConfig,
+		moduleFullName.Registry(),
+		container.Stdout(),
+	).PrintModule(ctx, format, modules[0])
 }